ops/legacy/Jarvis-Desktop-Agent/requirements.txt
mss
Pillow
pytesseract
pyautogui
